Update custom protocol widget
Now it supports Union[float, str] parameters, e.g. those used in CC/CV.
`default_value` added to CyclingParameter schema.
`value` will be the actual selected value. If it is not selected, it
should be None.
The TechniqueParametersWidget has been updated: each optional parameter
can be activated with a checkbox. If it is not activated, `value` will
be set to None.