Make all cycling params with defaults required=`True`
In the current instrument, cycling parameters for many techniques have
default values. These are not strictly required: if they are not given,
the `default_value` will be used.
To make the `CyclingTechnique` data more clear and complete, we set all
these parameters as required, so their value will be set.
(Reminder: all parameters values that are None will not be included in
the payload. For now these will be just the ones with
`required == False and default_value == None`)