Add Results Visualizer app
"Aurora app" now implements only submit functions.
The "Results Visualizer" has some basic functionalities to show the
list of submitted jobs (monitored jobs are not visualized, for
simplicity). The user can select a job and visualize the results.
For now, only these plots are implemented:
- voltage vs time
- current vs time
- voltage & current vs time
- capacity vs cycle number

BUG - TODO: a new matplotlib figure is generated every time the plot
button is clicked and the figure is appended. Ideally one should update
the same figure. Moreover, I was not able to clear the output once
generated, not sure why.